refactor: revamped folder structure of mongol db, shouldve done it in multiple commits
