+++ src/db/mongol.rs
@@ -1,9 +1,19 @@
-mod model;
-mod repositories;
-mod helper;
+pub mod helper;
+mod bucket;
+mod chat;
+mod message;
+mod refresh_token;
+mod relation;
+mod server;
+mod user;
 
-pub use model::*;
-pub use helper::*;
+pub use bucket::*;
+pub use chat::*;
+pub use message::*;
+pub use refresh_token::*;
+pub use relation::*;
+pub use server::*;
+pub use user::*;
 
 use std::time::Duration;
 use mongodb::{options::{ClientOptions, Compressor}, Client, Collection};
