--- src/db/mongol/model.rs
@@ -1,17 +1,0 @@
-mod user;
-mod chat_info; 
-mod message; 
-mod bucket; 
-mod refresh_token;
-mod relation;
-mod chat;
-mod server;
-
-pub use user::*;
-pub use chat_info::*;
-pub use message::*;
-pub use bucket::*; 
-pub use refresh_token::*;
-pub use relation::*;
-pub use chat::*;
-pub use server::*;
