+++ src/db/mongol/refresh_token.rs
@@ -1,8 +1,11 @@
+mod repository;
+
+
 use bson::{Bson, DateTime, Uuid};
 use serde::{Deserialize, Serialize};
 
 use crate::model::{error, refresh_token::{self, RefreshToken}};
-use crate::db::mongol::{as_string, helper, MongolHelper};
+use super::helper::{self, as_string, MongolHelper};
 
 
 #[derive(Debug, Serialize, Deserialize)]
