+++ src/db/mongol/relation.rs
@@ -1,3 +1,6 @@
+mod repository;
+
+
 use bson::Uuid;
 use serde::{Deserialize, Serialize};
