--- src/db/mongol/repositories.rs
@@ -1,6 +1,0 @@
-mod user;
-mod chat;
-mod message;
-mod refresh_token;
-mod relation;
-mod server;
